packages/python/examples/rag/requirements.txt
sentence-transformers
llama-index
tiktoken
numpy
logging
pathlib
pyyaml
dynaconf
dashscope
zhipuai
openai
random
docx
